Fix a NULL pointer issue in request.c
With a malformed request line, rent.method would not be populated,
and thus would remain NULL. This caused problems with how we determine
outputting the body of a page, as strcmp() was used to see whether or
not the request method wanted a body. If rent.method had the value
"HEAD", we would ignore outputting it. Since strcmp(), however, is not
NULL-safe, this would lead to a segmentation fault in the request
process.